Factor out intgration test createSuiteId() to trait
