+++ tests/Integration/MachineCreateMessageHandlingTest.php
@@ -9,14 +9,11 @@ use App\Model\RemoteRequestType;
 use App\Repository\RemoteRequestRepository;
 use App\Tests\Application\AbstractCreateJobSuccessSetup;
 use App\Tests\Services\EntityRemover;
-use App\Tests\Services\Generator\StringValue;
-use SmartAssert\TestSourcesClient\FileClient;
-use SmartAssert\TestSourcesClient\FileSourceClient;
-use SmartAssert\TestSourcesClient\SuiteClient;
 
 class MachineCreateMessageHandlingTest extends AbstractCreateJobSuccessSetup
 {
     use GetClientAdapterTrait;
+    use CreateSuiteIdTrait;
 
     private const int MICROSECONDS_PER_SECOND = 1000000;
 
@@ -57,32 +54,6 @@ class MachineCreateMessageHandlingTest extends AbstractCreateJobSuccessSetup
             $successfulMachineCreateRequestCount,
             'Incorrect machine/create request count, should be only one.'
         );
-    }
-
-    protected static function createSuiteId(): string
-    {
-        $fileSourceClient = self::getContainer()->get(FileSourceClient::class);
-        \assert($fileSourceClient instanceof FileSourceClient);
-
-        $fileSourceLabel = StringValue::random();
-
-        $fileSourceId = $fileSourceClient->create(self::$apiToken, $fileSourceLabel);
-        \assert(is_string($fileSourceId));
-
-        $fileClient = self::getContainer()->get(FileClient::class);
-        \assert($fileClient instanceof FileClient);
-
-        $fileClient->add(self::$apiToken, $fileSourceId, 'test.yaml', '- test file content');
-
-        $suiteClient = self::getContainer()->get(SuiteClient::class);
-        \assert($suiteClient instanceof SuiteClient);
-
-        $suiteLabel = StringValue::random();
-
-        $suiteId = $suiteClient->create(self::$apiToken, $fileSourceId, $suiteLabel, ['test.yaml']);
-        \assert(is_string($suiteId));
-
-        return $suiteId;
     }
 
     private function waitUntilSuccessfulMachineCreateRemoteRequestExists(string $jobId): void
