+++ tests/Integration/SerializedSuiteIsPreparedBeforeExplicitlyRetrieved.php
@@ -9,15 +9,12 @@ use App\Repository\RemoteRequestRepository;
 use App\Repository\SerializedSuiteRepository;
 use App\Tests\Application\AbstractCreateJobSuccessSetup;
 use App\Tests\Services\EntityRemover;
-use App\Tests\Services\Generator\StringValue;
 use Doctrine\ORM\EntityManagerInterface;
-use SmartAssert\TestSourcesClient\FileClient;
-use SmartAssert\TestSourcesClient\FileSourceClient;
-use SmartAssert\TestSourcesClient\SuiteClient;
 
 class SerializedSuiteIsPreparedBeforeExplicitlyRetrieved extends AbstractCreateJobSuccessSetup
 {
     use GetClientAdapterTrait;
+    use CreateSuiteIdTrait;
 
     private const int MICROSECONDS_PER_SECOND = 1000000;
 
@@ -49,32 +46,6 @@ class SerializedSuiteIsPreparedBeforeExplicitlyRetrieved extends AbstractCreateJ
         ]);
 
         self::assertNull($serializedSuiteRetrieveRequest);
-    }
-
-    protected static function createSuiteId(): string
-    {
-        $fileSourceClient = self::getContainer()->get(FileSourceClient::class);
-        \assert($fileSourceClient instanceof FileSourceClient);
-
-        $fileSourceLabel = StringValue::random();
-
-        $fileSourceId = $fileSourceClient->create(self::$apiToken, $fileSourceLabel);
-        \assert(is_string($fileSourceId));
-
-        $fileClient = self::getContainer()->get(FileClient::class);
-        \assert($fileClient instanceof FileClient);
-
-        $fileClient->add(self::$apiToken, $fileSourceId, 'test.yaml', '- test file content');
-
-        $suiteClient = self::getContainer()->get(SuiteClient::class);
-        \assert($suiteClient instanceof SuiteClient);
-
-        $suiteLabel = StringValue::random();
-
-        $suiteId = $suiteClient->create(self::$apiToken, $fileSourceId, $suiteLabel, ['test.yaml']);
-        \assert(is_string($suiteId));
-
-        return $suiteId;
     }
 
     private function waitUntilSerializedSuiteStateIsPrepared(int $waitThresholdInSeconds): void
